requirements.txt
python-dotenv
python-env
semantic_kernel
asyncio
pandas
pillow
requests
azure.ai.projects
pyodbc
fastapi
uvicorn[standard]
pydantic
azure-identity

azure.ai.projects
azure.ai.agents
azure.identity
pandas

# Required by xlxs to md
openpyxl
tabulate
IPython


pyyaml
opentelemetry-sdk
azure-mgmt-monitor